% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_method-9_COV.R
\name{DSPNorm.eval.cov}
\alias{DSPNorm.eval.cov}
\title{Evaluate Performance of Normlaization Methods by Comparing Coefficients of Variation}
\usage{
DSPNorm.eval.cov(normed.data.list, norm.methods.list, controls, main = "")
}
\arguments{
\item{normed.data.list}{a list object containing the normalized DSP expression matrix under all normalization methods, each normalized matrix can be either a data.frame or matrix.}

\item{norm.methods.list}{a numerical vector of the method IDs to be compared, for instance, norm.methods.list = 1:17.}

\item{controls}{a character vector listing both housekeeping and negative control IDs, which are to be ignored in evaluation of COV.}

\item{main}{title to be added to the output figure.}
}
\value{
ggplot2 object of box plot.
}
\description{
The function generates box plot of coefficients of variation (COV) of protein across the ROIs of biological replicates. A better normalization is expected to yield lower level of COV.
}
\details{
To add some details here!
}
\examples{
data(COVID.19.DSP)

data.normed.list <- COVID.19.DSP$Expr.normed.DSP

group.info <- COVID.19.DSP$Meta.DSP

controls <- c("Ms IgG1", "Ms IgG2a", "Rb IgG",
              "S6", "GAPDH", "Histone H3")

pick.id.1 <- group.info$organ.short=="Alveolar"
pick.id.2 <- group.info$PanCK.Status=="PanCK+"
pick.id.3 <- group.info$X.Sample_characteristics_ch1==
                "infection: SARS-CoV-2 infected"

norm.data.list.s <- list()
for (method_id in c(1:5, 8:16)) {
  data.pick <- data.normed.list[[method_id]]
  norm.data.list.s[[method_id]] <-
     data.pick[ ,pick.id.1 & pick.id.2 & pick.id.3]
}

DSPNorm.eval.cov(normed.data.list = norm.data.list.s,
                 norm.methods.list = c(1:5, 8:16),
                 controls = controls,
                 main = "All Patients/Alveolar/Infected/PanCK+")
}
