% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_method-3_heatmap.R
\name{DSPNorm.eval.heatmap}
\alias{DSPNorm.eval.heatmap}
\title{Evaluate Performance of Normlaization Methods by Comparing clustering based on Euclidean distance.}
\usage{
DSPNorm.eval.heatmap(
  normed.data.list,
  norm.methods.list,
  grouping.info,
  main = "",
  title.size = 7
)
}
\arguments{
\item{normed.data.list}{a list object containing the normalized DSP expression matrix under all normalization methods, each normalized matrix can be either a data.frame or matrix.}

\item{norm.methods.list}{a numerical vector of the method IDs to be compared, for instance, norm.methods.list = 1:17.}

\item{grouping.info}{a character vector describing grouping info of each ROI, must be arranged in the same order of the columns of DSP expression matrix. Each unique character in the vector is to be treated as a unique biological subgroup. For optimal visualization, it is strongly recommended to consider only two unique biological groups of interest in each round of comparison.}

\item{main}{a character string. If unspecified, then the method numbers are displayed as title. If specified, then the specified string will be added in front of the method numbers in title.}

\item{title.size}{font size of title.}
}
\value{
a pheatmap object.
}
\description{
The function generates Euclidean-based clustering plots basded on normalized DSP data. A better normalization is expected to yield a clustering prediction concistent with known biological subgroups.
}
\examples{
data(COVID.19.DSP)

data.normed.list <- COVID.19.DSP$Expr.normed.DSP

group.info <- COVID.19.DSP$Meta.DSP

controls <- c("Ms IgG1", "Ms IgG2a", "Rb IgG",
              "S6", "GAPDH", "Histone H3")

group.info$PanCK.Status[group.info$PanCK.Status=="PanCK-"] <- "PanCK Neg"
group.info$PanCK.Status[group.info$PanCK.Status=="PanCK+"] <- "PanCK Pos"

pick.id.1 <- group.info$organ.short=="Alveolar"

pick.id.2 <- group.info$patient=="D9"

group.info.pick <- group.info[pick.id.1 & pick.id.2, ]

norm.data.list.s <- list()
for (method_id in c(1:5, 8:16)) {
  data.pick <- log(data.normed.list[[method_id]])
  norm.data.list.s[[method_id]] <-
     data.pick[!row.names(data.pick)\%in\%controls, pick.id.1 & pick.id.2]
}

DSPNorm.eval.heatmap(normed.data.list = norm.data.list.s,
                     norm.methods.list = c(1:5, 8:16),
                     grouping.info = group.info.pick$PanCK.Status,
                     main = "Sample D9",
                     title.size = 7.5)
}
