% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_method-1_tech-var-minimize.R
\name{DSPNorm.eval.tech.var}
\alias{DSPNorm.eval.tech.var}
\title{Evaluate Performance of Normlaization Methods by Comparing Technical Variation among Replicates}
\usage{
DSPNorm.eval.tech.var(
  normed.data.list,
  norm.methods.list,
  grouping.info,
  main = "",
  title.size = 6,
  deviation.from.median = FALSE,
  group.focus = "treatment",
  range.universal = c(NA, NA)
)
}
\arguments{
\item{normed.data.list}{a list object containing the normalized DSP expression matrix under all normalization methods, each normalized matrix can be either a data.frame or matrix.}

\item{norm.methods.list}{a numerical vector of the method IDs to be compared, for instance, norm.methods.list = 1:17.}

\item{grouping.info}{a character vector describing grouping info of each ROI, must be arranged in the same order of the columns of DSP expression matrix.}

\item{main}{title to be added to the output figure.}

\item{title.size}{font size of title.}

\item{deviation.from.median}{if TRUE, the distribution of expressions of proteins within each ROI is evaluated relative to its median.}

\item{group.focus}{a character vector specifying which group in \code{"grouping.info"} is to be collected for distribution plot, the ROIs in the selected group is to be treated as biological replicates.}

\item{range.universal}{a numerical vector of length two, specifying the range of distribution to be plotted. If unspecified, then the range is decided automatically by the actual range of input data.}
}
\value{
ggplot2 object of distribution plot.
}
\description{
The function generates distribution plots of normalized DSP data of the ROIs that are treated as biological replicates by a list of normalization methods. A better normalization method is expected to yield smaller technical variations among biological replicates.
}
\examples{
data(COVID.19.DSP)

data.normed.list <- COVID.19.DSP$Expr.normed.DSP

group.info <- COVID.19.DSP$Meta.DSP

controls <- c("Ms IgG1", "Ms IgG2a", "Rb IgG",
              "S6", "GAPDH", "Histone H3")

pick.id.1 <- group.info$organ.short=="Alveolar"

pick.id.2 <- group.info$patient=="D9"

group.info.pick <- group.info[pick.id.1 & pick.id.2, ]

norm.data.list.s <- list()
for (method_id in c(1:5, 8:16)) {
  data.pick <- log(data.normed.list[[method_id]])
  norm.data.list.s[[method_id]] <-
     data.pick[!row.names(data.pick)\%in\%controls, pick.id.1 & pick.id.2]
}

DSPNorm.eval.tech.var(normed.data.list = norm.data.list.s,
                      norm.methods.list = c(1:5, 8:16),
                      grouping.info = group.info.pick[ ,"PanCK.Status"],
                      main = "Sample D9",
                      title.size = 6.5,
                      group.focus = "PanCK+",
                      range.universal = c(-2.5,9.5))
}
