% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSP_Norm.R
\name{DSPNorm}
\alias{DSPNorm}
\title{Normalize Data of NanoString GeoMx Protein Assays}
\usage{
DSPNorm(
  data.raw = NULL,
  control.AreaToPixelSquare = NULL,
  control.Nuclei = NULL,
  control.HK = NULL,
  control.SignalToNoise = NULL,
  normalization.method = "Area_geoMean",
  control.rm = T
)
}
\arguments{
\item{data.raw}{a raw DSP data matrix that has only been ERCC-normalized. The raw data matrix may be either a matrix or data.frame object with each row for a protein and each column for a ROI.}

\item{control.AreaToPixelSquare}{a numerical vector of ROI areas, must be arranged in the same order of the columns of data.raw. If missing, the normalization methods concerning area cannot be performed.}

\item{control.Nuclei}{a numerical vector of ROI nuclei counts, must be arranged in the same order of the columns of data.raw. If missing, the normalization methods concerning nuclei count cannot be performed.}

\item{control.HK}{a character vector listing all the housekeeping proteins that appear in the row names of data.raw.}

\item{control.SignalToNoise}{a character vector listing all the negative control proteins that appear in the row names of data.raw.}

\item{normalization.method}{a character specifying which normalization method is applied to normalize raw DSP data. Available normalization methods include: "None", "Area_geoMean", "Nuclei_geoMean", "STNR_geoMean", "STNS_mean2SD", "HK_geoMean".}

\item{control.rm}{a logical value indicating whether the housekeep/negative controls should be removed from the output of normalized data matrix.}
}
\value{
a matrix or data.frame object, same class and data structure as the input of raw DSP data matrix.
}
\description{
NanostringNorm provides flexible options of normalization methods for raw data (only ERCC-normalized) of NanoString GeoMx protein assays for downstream performance evaluation.

Note that certain normalization methods may be combined in a serial manner to constitute an addtional normalization method that is not on the list of available options in the 'normalization.method' parameter.
}
\details{
To add some details here!
}
