% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboxplot.R
\name{myboxplot.DSP}
\alias{myboxplot.DSP}
\title{For internal usage of box plot}
\usage{
myboxplot.DSP(
  y,
  x,
  title = "",
  type = c("Group A", "Group B"),
  ylab = "Protein Expr"
)
}
\arguments{
\item{y}{vector of values to plot}

\item{x}{vector of grouping info}

\item{title}{title to show}

\item{type}{groups of interest}

\item{ylab}{ylab to show}
}
\description{
For internal usage of box plot
}
