% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_method-1_bio-var-maximize.R
\name{norm.eval.bio.var}
\alias{norm.eval.bio.var}
\title{Evaluate Performance of Normlaization Methods by Comparing Biological Variation}
\usage{
norm.eval.bio.var(
  normed.data.list,
  norm.methods.list,
  grouping.info,
  main = "",
  title.size = 6
)
}
\arguments{
\item{normed.data.list}{a list object containing the normalized DSP expression matrix under all normalization methods, each normalized matrix can be either a data.frame or matrix.}

\item{norm.methods.list}{a numerical vector of the method IDs to be compared, for instance, norm.methods.list = 1:17.}

\item{grouping.info}{a character vector describing grouping info of each ROI, must be arranged in the same order of the columns of DSP expression matrix. Each unique character in the vector is to be treated as a unique biological subgroup. For optimal visualization, it is strongly recommended to consider only two unique biological groups of interest in each round of comparison.}

\item{main}{title to be added to the output figure.}

\item{title.size}{font size of title.}
}
\value{
ggplot2 object of PCA plot.
}
\description{
The function generates PCA plots of normalized DSP data of the ROIs that are treated as biological replicates by a list of normalization methods. A better normalization method is expected to yield larger biological variations between two known biological subgroups.
}
\details{
To add some details here!
}
