% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_method-9_COV.R
\name{norm.eval.cov}
\alias{norm.eval.cov}
\title{Evaluate Performance of Normlaization Methods by Comparing Coefficients of Variation}
\usage{
norm.eval.cov(normed.data.list, norm.methods.list, controls, main = "")
}
\arguments{
\item{normed.data.list}{a list object containing the normalized DSP expression matrix under all normalization methods, each normalized matrix can be either a data.frame or matrix.}

\item{norm.methods.list}{a numerical vector of the method IDs to be compared, for instance, norm.methods.list = 1:17.}

\item{controls}{a character vector listing both housekeeping and negative control IDs, which are to be ignored in evaluation of COV.}

\item{main}{title to be added to the output figure.}
}
\value{
ggplot2 object of box plot.
}
\description{
The function generates box plot of coefficients of variation (COV) of protein across the ROIs of biological replicates. A better normalization is expected to yield lower level of COV.
}
\details{
To add some details here!
}
