% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_method-3_heatmap.R
\name{norm.eval.heatmap}
\alias{norm.eval.heatmap}
\title{Evaluate Performance of Normlaization Methods by Comparing clustering based on Euclidean distance.}
\usage{
norm.eval.heatmap(
  normed.data.list,
  norm.methods.list,
  grouping.info,
  main = "",
  title.size = 7
)
}
\arguments{
\item{normed.data.list}{a list object containing the normalized DSP expression matrix under all normalization methods, each normalized matrix can be either a data.frame or matrix.}

\item{norm.methods.list}{a numerical vector of the method IDs to be compared, for instance, norm.methods.list = 1:17.}

\item{grouping.info}{a character vector describing grouping info of each ROI, must be arranged in the same order of the columns of DSP expression matrix. Each unique character in the vector is to be treated as a unique biological subgroup. For optimal visualization, it is strongly recommended to consider only two unique biological groups of interest in each round of comparison.}

\item{main}{title to be added to the output figure.}

\item{title.size}{font size of title.}
}
\value{
a pheatmap object.
}
\description{
The function generates Euclidean-based clustering plots basded on normalized DSP data. A better normalization is expected to yield a clustering prediction concistent with known biological subgroups.
}
\details{
To add some details here!
}
