% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_method-2_silhouette_width.R
\name{norm.eval.silhouette.width}
\alias{norm.eval.silhouette.width}
\title{Evaluate Performance of Normlaization Methods by Comparing silhouette width}
\usage{
norm.eval.silhouette.width(
  normed.data.list,
  norm.methods.list,
  grouping.info,
  main = "",
  title.size = 7,
  return.values.only = F
)
}
\arguments{
\item{normed.data.list}{a list object containing the normalized DSP expression matrix under all normalization methods, each normalized matrix can be either a data.frame or matrix.}

\item{norm.methods.list}{a numerical vector of the method IDs to be compared, for instance, norm.methods.list = 1:17.}

\item{grouping.info}{a character vector describing grouping info of each ROI, must be arranged in the same order of the columns of DSP expression matrix. Each unique character in the vector is to be treated as a unique biological subgroup.}

\item{main}{title to be added to the output figure.}

\item{title.size}{font size of title.}

\item{return.values.only}{if TRUE, then only numerical table of computed silhouett widths, instead of violin plot, will be returned.}
}
\value{
ggplot2 object of violin plot. A numerical table of data.frame class will be returned if \code{"return.values.only"} is set as TRUE.
}
\description{
The function generates violin plot of silhouette widths basded on normalized DSP data and known biological groups of interest each ROI belongs to. A better normalization is expected to yield higher values of silhouette width, which indicate higher chance of correctly classifying each ROI into the biological group it actually belongs to.
}
\details{
To add some details here!
}
