% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_method-5_spearman_correlation.R
\name{norm.eval.spearman.correlation}
\alias{norm.eval.spearman.correlation}
\title{Evaluate Performance of Normlaization Methods by Comparing Spearman correlation between DSP protein expression and the corresponding mRNA-level expression, if available}
\usage{
norm.eval.spearman.correlation(
  normed.data.list,
  norm.methods.list,
  reference.data.matrix,
  main = ""
)
}
\arguments{
\item{normed.data.list}{a list object containing the normalized DSP expression matrix under all normalization methods, each normalized matrix can be either a data.frame or matrix.}

\item{norm.methods.list}{a numerical vector of the method IDs to be compared, for instance, norm.methods.list = 1:17.}

\item{reference.data.matrix}{mRNA-Seq expression data of either a matrix or a data.frame class. Must be arranged in the exactly the same order of the DSP protein expression data.}

\item{main}{title to be added to the output figure.}
}
\value{
ggplot2 object of box plot, for both Spearman correlation strengths and Spearman correlation p values.
}
\description{
The function generates box plot of Spearman correlations of matched protein-mRNA expressions. A better normalization method is expected to yield stronger correlation between protein expression and the correponding mRNA expression within the same ROI.
}
\details{
To add some details here!
}
