% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_method-1_tech-var-minimize.R
\name{norm.eval.tech.var}
\alias{norm.eval.tech.var}
\title{Evaluate Performance of Normlaization Methods by Comparing Technical Variation among Replicates}
\usage{
norm.eval.tech.var(
  normed.data.list,
  norm.methods.list,
  grouping.info,
  main = "",
  title.size = 6,
  deviation.from.median = F,
  group.focus = "treatment",
  range.universal = c(NA, NA)
)
}
\arguments{
\item{normed.data.list}{a list object containing the normalized DSP expression matrix under all normalization methods, each normalized matrix can be either a data.frame or matrix.}

\item{norm.methods.list}{a numerical vector of the method IDs to be compared, for instance, norm.methods.list = 1:17.}

\item{grouping.info}{a character vector describing grouping info of each ROI, must be arranged in the same order of the columns of DSP expression matrix.}

\item{main}{title to be added to the output figure.}

\item{title.size}{font size of title.}

\item{deviation.from.median}{if TRUE, the distribution of expressions of proteins within each ROI is evaluated relative to its median.}

\item{group.focus}{a character vector specifying which group in \code{"grouping.info"} is to be collected for distribution plot, the ROIs in the selected group is to be treated as biological replicates.}

\item{range.universal}{a numerical vector of length two, specifying the range of distribution to be plotted. If unspecified, then the range is decided automatically by the actual range of input data.}
}
\value{
ggplot2 object of distribution plot.
}
\description{
The function generates distribution plots of normalized DSP data of the ROIs that are treated as biological replicates by a list of normalization methods. A better normalization method is expected to yield smaller technical variations among biological replicates.
}
\details{
To add some details here!
}
