% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSP_QC.R
\name{QC.DSP}
\alias{QC.DSP}
\title{Check Quality of Data of NanoString GeoMx Protein Assays}
\usage{
QC.DSP(
  data.raw,
  Controls,
  colData,
  Comparison.Pair,
  colData.grouping.index,
  colData.ROI.area.index,
  colData.ROI.nuclei.count.index = NULL,
  colData.Subject.index = NULL
)
}
\arguments{
\item{data.raw}{a raw DSP data matrix that has only been ERCC-normalized. The raw data matrix may be either a matrix or data.frame object with each row for a protein and each column for a ROI.}

\item{Controls}{a character vector listing both the negative control and housekeeping proteins that appear in the row names of data.raw, preferably in the order of negative controls first followed by housekeepers.}

\item{colData}{a data.frame object providing annotation info associated with the DSP data, such as area/nuclei count of each ROI, grouping info of each ROI, and subject to which each ROI belongs. The rows of colData and the columns of data.raw must be arranged in exactly the same order.}

\item{Comparison.Pair}{a character vector of length two, which can be found in the grouping info column of colData, as specified by the argument colData.grouping.index.}

\item{colData.grouping.index}{a character specifying the name of the column of colData containing the grouping info of the study.}

\item{colData.ROI.area.index}{a character specifying the name of the column of colData containing the area info of ROIs.}

\item{colData.ROI.nuclei.count.index}{a character specifying the name of the column of colData containing the nuclei count info of ROIs. If missing, then the QC steps concerning nuclei count are skipped.}

\item{colData.Subject.index}{a character specifying the name of the column of colData containing the subject info of ROIs. if missing, then single subject is assumed.}
}
\value{
graphs for QC diagnosis.
}
\description{
QC.DSP provides several measures to check quality of NanoString GeoMx Protein DSP data, revealing potential data quality issues before proceeding to downstream analysis, also providing important clues in deciding optimal normalization methods.
}
\examples{
data(COVID.19.DSP)

data.raw <- COVID.19.DSP$Expr.raw.DSP

group.info <- COVID.19.DSP$Meta.DSP

control.SignalToNoise.use <- c("Ms IgG1", "Ms IgG2a", "Rb IgG")
control.HK.use <- c("Histone H3", "S6", "GAPDH")
control.both <- c(control.SignalToNoise.use, control.HK.use)

group.info.Alveolar <- group.info[group.info$organ.short=="Alveolar", ]

panck.uniq <- unique(group.info.Alveolar$PanCK.Status)

data.raw.Alveolar <- data.raw[ ,group.info$organ.short=="Alveolar"]

QC.DSP(data.raw = data.raw.Alveolar,
       Controls = control.both,
       colData = group.info.Alveolar,
       Comparison.Pair = panck.uniq,
       colData.grouping.index = "PanCK.Status",
       colData.ROI.area.index = "area",
       colData.Subject.index = "patient")
}
